pyproject.toml
[project]
name = "pixel-match"
version = "0.1.0"
description = "Image Search"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pydantic>=2.12.4",
    "numpy>=2.2.6", 
    "python-dotenv>=1.2.1", 
    "pillow>=12.0.0", 
    "faiss-cpu>=1.12.0", 
    "pandas>=2.3.3", 
    "transformers>=4.57.1", 
    "streamlit>=1.51.0", 
    "torch>=2.9.0"
]

[project.optional-dependencies]
test = ["pytest>=9.0.1",]

[tool.pytest.ini_options]
pythonpath = "."
testpaths = ["tests"]

[tool.ruff.lint]
select = [
    "E", # pycodestyle errors
    "W", # pycodestyle warnings
    "F", # Pyflakes
    "I", # isort
    "UP", # pyupgrade
]

ignore = [
    "E501", # line-too-long
]